pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-gcptoolkit"
version = "0.1.0"
description = "GCP toolkit for agent operations"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-secret-manager",
]

[project.scripts]
gcptoolkit = "frontend.cli.main:main"

[tool.setuptools]
packages = [
    "agent_gcptoolkit",
    "backend",
    "backend.services",
    "backend.services.secrets",
    "backend.services.secrets.src",
    "backend.services.secrets.src.domains",
    "backend.services.secrets.src.workflows",
    "frontend",
    "frontend.cli",
]
